cloud_function_video_upload/requirements.txt
functions-framework==3.4.0
google-cloud-storage==2.9.0
google-cloud-aiplatform==1.35.0
protobuf==4.24.4
moviepy==1.0.3
